move all feature into freature directory
