+++ Makefile
@@ -3,6 +3,8 @@ PWD := $(shell pwd)
 all: run-backend run-database run-frontend
 
 dev:
+	npm install --prefix frontend/app
+	npm install --prefix backend
 	docker compose up --build --detach
 
 run-backend:
