+++ backend/src/features/banned/dto/return-ban.dto.ts
@@ -1,4 +1,3 @@
-import { ReturnChanelDto } from '@backend/channels/dto/return-cnannel.dto';
 import { User, ChannelsEntity } from '@backend/typeorm';
 import { IsNumber, IsString } from 'class-validator';
