+++ backend/src/features/features.module.ts
@@ -8,6 +8,10 @@ import { UserSessionModule } from './user-session/user-session.module';
 import { MatchsModule } from './matchs/matchs.module';
 import { GatewayModule } from '@backend/gateWay/gateway.module';
 import { StatsModule } from './stats/stats.module';
+import { ChannelsModule } from './channels/channels.module';
+import { MessgesModule } from './messages/messages.module';
+import { BannedModule } from './banned/banned.module';
+import { BlockUserModule } from './block/blockUser.module';
 
 @Module({
   imports: [
@@ -20,6 +24,10 @@ import { StatsModule } from './stats/stats.module';
     GatewayModule,
     MatchsModule,
     StatsModule,
+    ChannelsModule,
+    BannedModule,
+    MessgesModule,
+    BlockUserModule
   ],
   exports: [
     MeModule,
@@ -31,6 +39,10 @@ import { StatsModule } from './stats/stats.module';
     GatewayModule,
     MatchsModule,
     StatsModule,
+    ChannelsModule,
+    BannedModule,
+    MessgesModule,
+    BlockUserModule
   ],
 })
 export class FeaturesModule {}
