+++ backend/src/gateWay/chatSocket.module.ts
@@ -1,12 +1,11 @@
 import { Module } from '@nestjs/common';
 import { SocketGateway } from './chat.gateway';
 import { SocketService } from './chatSocket.service';
-// import { ChannelsService } from '@backend/channels/channels.service';
-import { ChannelsModule } from '@backend/channels/channels.module';
-import { BannedModule } from '@backend/banned/banned.module';
-import { MutedModule } from '@backend/muted/muted.module';
-import { MessgesModule } from '@backend/messages/messages.module';
 import { SharedModule } from '@backend/shared/shared.module';
+import { ChannelsModule } from '@backend/features/channels/channels.module';
+import { BannedModule } from '@backend/features/banned/banned.module';
+import { MutedModule } from '@backend/features/muted/muted.module';
+import { MessgesModule } from '@backend/features/messages/messages.module';
 
 @Module({
   providers: [SocketGateway, SocketService, ChannelsModule],
